requirements.txt
requests>=2.31
PyYAML>=6.0